[touch: 1045] fix: change field 'order' to integer

+++ database/migrations/2016_05_11_084328_create_event_types_table.php
@@ -16,7 +16,7 @@ class CreateEventTypesTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('icon');
-            $table->float('order')->nullable();
+            $table->integer('order')->nullable();
             $table->softDeletes()->nullable()->default(null);
             $table->timestamps();
         });
